move NewProductForm
