+++ client/src/pages/Admin.jsx
@@ -1,12 +1,11 @@
 import "../assets/css/admin.css";
-import { useState } from "react";
 import { Link } from "react-router-dom";
-// import exampleData from "../assets/example-data.json";
 import Auth from "../utils/auth";
 import { useMutation, useQuery } from "@apollo/client";
 import { GET_All_PRODUCTS } from "../utils/queries";
 import { toDecimal } from "../utils/math";
 import { ADD_PRODUCT } from "../utils/mutations";
+import NewProductForm from "../components/forms/NewProductForm";
 
 function AdminPage() {
   const { data } = useQuery(GET_All_PRODUCTS);
@@ -14,7 +13,7 @@ function AdminPage() {
   if (!Auth.isLoggedIn() || Auth.isClient() || !Auth.isAdmin()) {
     return <h1>you are not authorized to view this page</h1>;
   }
-  const [addProduct] = useMutation(ADD_PRODUCT);
+  // const [addProduct] = useMutation(ADD_PRODUCT);
   const { products } = data ? data : [];
   console.log(products);
 
@@ -50,114 +49,7 @@ function AdminPage() {
       </>
     );
   });
-  // TODO SEND INT TO DATABASE
-  const NewProductForm = () => {
-    const [formState, setFormState] = useState({
-      name: "",
-      price: "",
-      category: "",
-      description: "",
-      quantity: "",
-      image: undefined,
-      imageName: undefined,
-    });
-    const handleInputChange = (e) => {
-      const { name, value } = e.target;
 
-      setFormState({
-        ...formState,
-        [name]: value,
-      });
-    };
-
-    const handleFormSubmit = async (e) => {
-      e.preventDefault();
-
-      const { name, price, category, description, quantity } = formState;
-      console.log(formState);
-      try {
-        const { data } = addProduct({
-          variables: { name, price, category, description, quantity },
-        });
-        console.log(data);
-      } catch (err) {
-        console.log(err);
-      }
-    };
-
-    return (
-      <>
-        <form onSubmit={handleFormSubmit} className="new-product-form">
-          <h2>Add new product</h2>
-          <input
-            value={formState.name}
-            name="name"
-            onChange={handleInputChange}
-            type="text"
-            placeholder="Product name"
-          ></input>
-          {/* //TODO FORCE THIS INTO DECIMAL AND MAKE FUNCTION TO CONVERT INTO DATABASE SHTUFF */}
-          <input
-            value={formState.price}
-            name="price"
-            onChange={handleInputChange}
-            type="number"
-            placeholder="Price"
-            min="0"
-          ></input>
-          <input
-            value={formState.quantity}
-            name="quantity"
-            onChange={handleInputChange}
-            type="number"
-            min="0"
-            placeholder="Stock"
-          ></input>
-          {/* //TODO MAKE this a dropdown to limit CATEGORIES */}
-          <input
-            value={formState.category}
-            name="category"
-            onChange={handleInputChange}
-            type="text"
-            placeholder="Category"
-          ></input>
-          <textarea
-            value={formState.description}
-            name="description"
-            onChange={handleInputChange}
-            type="text"
-            placeholder="Description"
-          ></textarea>
-          <input
-            value={formState.image}
-            name="image"
-            onChange={handleInputChange}
-            type="file"
-          ></input>
-          <input
-            value={formState.imageName}
-            name="imageName"
-            onChange={handleInputChange}
-            type="text"
-            placeholder="Image title"
-          ></input>
-          <input
-            value={formState.imageDescription}
-            name="imageDescription"
-            onChange={handleInputChange}
-            type="text"
-            placeholder="Image caption"
-          ></input>
-
-          <button className="btn-1" type="submit">
-            Submit
-          </button>
-
-          <div className="form-footer"></div>
-        </form>
-      </>
-    );
-  };
 
   return (
     <>
